requirements.txt
anyio==4.13.0
arrow==1.4.0
binaryornot==0.4.4
certifi==2026.2.25
chardet==5.2.0
charset-normalizer==3.4.7
click==8.3.2
cookiecutter==2.7.1
flet==0.84.0
flet-cli==0.84.0
flet-desktop==0.84.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.1
psutil==7.2.2
Pygments==2.20.0
PyQt6==6.11.0
PyQt6-Qt6==6.11.0
PyQt6_sip==13.11.1
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-slugify==8.0.4
PyYAML==6.0.3
qrcode==8.2
repath==0.9.0
requests==2.33.1
rich==15.0.0
scapy==2.7.0
six==1.17.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
